telcox_backend/requirements.txt
asgiref==3.9.2
Django==5.2.7
django-cors-headers==4.9.0
djangorestframework==3.16.1
mysqlclient==2.2.7
python-decouple==3.8
sqlparse==0.5.3
typing_extensions==4.15.0
